+++ src/main/exec.c
@@ -152,8 +152,8 @@ pid_t radius_start_program(char const *cmd, REQUEST *request, bool exec_wait,
 				for (p = buffer; *p != '='; p++) {
 					if (*p == '-') {
 						*p = '_';
-					} else if (isalpha((int) *p)) {
-						*p = toupper(*p);
+					} else if (isalpha((uint8_t) *p)) {
+						*p = toupper((uint8_t) *p);
 					}
 				}
 			}
